Choose random clients
Additionally: Prices are INTEGERS

+++ src/Dish.java
@@ -1,11 +1,11 @@
 
 public class Dish extends MenuItem
 {
-    private static final double LOW_INGREDIENT_COST = 3;
-    private static final double HIGH_INGREDIENT_COST = 10;
+    private static final Integer LOW_INGREDIENT_COST = 3;
+    private static final Integer HIGH_INGREDIENT_COST = 10;
     
-    public static double lowQualityPrice = LOW_INGREDIENT_COST;
-    public static double highQualityPrice = HIGH_INGREDIENT_COST;
+    public static Integer lowQualityPrice = LOW_INGREDIENT_COST;
+    public static Integer highQualityPrice = HIGH_INGREDIENT_COST;
 	private Integer calorieCount;
 
 	public Dish( String name, Integer calorieCount ) {
@@ -15,7 +15,7 @@ public class Dish extends MenuItem
 	}
 	
 	@Override
-	public double calculateIngredientCost( ) {
+	public Integer calculateIngredientCost( ) {
 		if(qualityLevel.equals(QualityLevel.HIGH)){
 			return HIGH_INGREDIENT_COST;
 		}else{
@@ -24,7 +24,7 @@ public class Dish extends MenuItem
 	}
 
 	@Override
-	public double getPrice( ) {
+	public Integer getPrice( ) {
 		if(qualityLevel.equals(QualityLevel.LOW)){
 			return lowQualityPrice;
 		}else{
